Further Refactoring (#32)
