+++ kosat-core/src/jvmTest/kotlin/org/kosat/DiamondTests.kt
@@ -63,19 +63,15 @@ internal class DiamondTests {
         }
     }
 
-    private fun checkKoSatSolution(ans: Model, input: String, isSolution: Boolean): Boolean {
-        val values = ans.values ?: return !isSolution
+    private fun checkKoSatSolution(ans: List<Boolean>?, input: String, isSolution: Boolean): Boolean {
+        if (ans == null) return !isSolution
 
         val cnfRequest = readCnfRequests(input).first()
-        if (values.size != cnfRequest.vars) return false
+        if (ans.size != cnfRequest.vars) return false
 
         return cnfRequest.clauses.all { clause ->
             clause.toClause().any {
-                if (it.isPos) {
-                    values[it.variable] == LBool.TRUE
-                } else {
-                    values[it.variable] == LBool.FALSE
-                }
+                it.isPos == ans[it.variable]
             }
         }
     }
@@ -149,6 +145,13 @@ internal class DiamondTests {
 
             val (solution, timeKoSat) = measureTimeWithResult {
                 solver.solve(assumptions.map { it.toLiteral() })
+                val result = solver.solve(assumptions.map { it.toLiteral() })
+                if (result == SolveResult.SAT) {
+                    println(solver.getModel())
+                    solver.getModel()
+                } else {
+                    null
+                }
             }
 
             val checkRes = if (checkKoSatSolution(solution, input, isSolution)) "OK" else "WA"
